refactor(export-multiple-campaigns): implement as chunk task
